backend/requirements.txt
Django==5.1
djangorestframework==3.15
djangorestframework-simplejwt==5.4
channels==4.1
django-cors-headers==4.4
django-filter==24.3
channels-redis==4.2
celery==5.4
redis==5.1
psycopg2-binary==2.9.10
paho-mqtt==2.1
daphne==4.1
gunicorn==23.0
python-decouple==3.8
whitenoise==6.7
factory-boy==3.3
pytest-django==4.9
pytest-asyncio==0.24
pytest==8.3
